requirements.txt
flask==2.0.1
pillow==9.0.0
gunicorn==20.1.0